+++ release-notes.md
@@ -1,3 +1,8 @@
+## 2020-07-27
+- License added.
+- /time endpoint added, to get the current time on devices without an RTC.
+- /install/1 and /install/2 endpoints added, to get started with the EcoBee API.
+- Started refactoring HTTP functions into reusable functions.
 ## 2020-07-22
 - Fix opening lots of connections to the DB.
 ## 2020-07-21